fixing smells and bugs

+++ src/main/java/nl/tudelft/jpacman/npc/Ghost.java
@@ -74,7 +74,7 @@ public abstract class Ghost extends Unit {
      * @return The suggested delay between moves in milliseconds.
      */
     public long getInterval() {
-        return this.moveInterval + new SecureRandom().nextInt(this.intervalVariation);
+        return this.moveInterval + (long) new SecureRandom().nextInt(this.intervalVariation);
     }
 
     /**
